backend/requirements.txt
# FastAPI Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Web Scraping
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3

# LLM and AI
langchain==0.1.0
langchain-google-genai==0.0.6
google-generativeai==0.3.1

# Environment Variables
python-dotenv==1.0.0

# CORS
python-multipart==0.0.6

# Date and Time
python-dateutil==2.8.2

# JSON handling
pydantic==2.5.0
pydantic-settings==2.1.0